Turn type-based constants into runtime-based ones
This is necessary if we want to have a `Kernel` that's actually just a
union of mulitple possible kernels and which one is actually used is
determined at runtime.

For example, one can imagine an enum like so:

```rust
pub enum AllowedKernels {
    Matern1(friedrich::kernel::Matern1),
    Matern2(friedrich::kernel::Matern2)
}
```

and at runtime we specify which kernel to use or even change the kernel
mid-way if it's underperforming and retrain.

Without the change in the commit however, it's impossible to implement
the `Kernel` trait: we don't know ahead of time which kernel will be
used so `NB_PARAMETERS` and `IS_SCALEABLE` are not possible to implement
for unions.

As `NB_PARAMETERS` and `IS_SCALEABLE` are never used in a fashion that
actually exploits their compile-time-constness, this seems like an
acceptable change. I appreciate that we can imagine a world where we use
`NB_PARAMETERS` as an array length or that we will get rid of
`IS_SCALEABLE` in favour of a trait that stops invalid usage at compile
time. Both these changes make unions impossible again.